moved editor as scaffold from page
