+++ editor/scaffolds/README.md
@@ -1,0 +1,1 @@
+# Stetefull + logic + ui scaffolds.
